Remove database configuration from installer script
It is not very useful to change this stuff and if one wants, it can be 
done manually in the env file (which is also stated at the end of the 
script). An argument can be made to set a new password, but since the DB 
cannot be accessed from outside the risk should be minimal.